[DOCKER] Consolidate CLI definition in the `cli` module
This is mainly to consolidate related code, and lock down visibility
as a prelude to further cleanups.
