+++ components/pkg-export-docker/src/cli.rs
@@ -1,6 +1,7 @@
 use crate::RegistryType;
 use clap::{App,
            Arg};
+use habitat_common::PROGRAM_NAME;
 use habitat_core::package::PackageIdent;
 use std::{path::Path,
           result,
@@ -8,11 +9,29 @@ use std::{path::Path,
 use url::Url;
 
 /// The version of this library and program when built.
-pub const VERSION: &str = include_str!(concat!(env!("OUT_DIR"), "/VERSION"));
+const VERSION: &str = include_str!(concat!(env!("OUT_DIR"), "/VERSION"));
+
+/// Create the Clap CLI for the Docker exporter
+pub fn cli<'a, 'b>() -> App<'a, 'b> {
+    let name: &str = &*PROGRAM_NAME;
+    let about = "Creates (and optionally pushes) a Docker image from a set of Habitat packages";
+
+    let mut cli = Cli::new(name, about).add_base_packages_args()
+                                       .add_builder_args()
+                                       .add_tagging_args()
+                                       .add_publishing_args()
+                                       .add_memory_arg()
+                                       .add_layer_arg()
+                                       .add_pkg_ident_arg(PkgIdentArgOptions { multiple: true });
+    if cfg!(windows) {
+        cli = cli.add_base_image_arg();
+    }
+    cli.app
+}
 
 /// A Docker-specific clap:App wrapper
 #[derive(Clone)]
-pub struct Cli<'a, 'b>
+struct Cli<'a, 'b>
     where 'a: 'b
 {
     pub app: App<'a, 'b>,
@@ -24,7 +43,7 @@ pub struct PkgIdentArgOptions {
 }
 
 impl<'a, 'b> Cli<'a, 'b> {
-    pub fn new(name: &str, about: &'a str) -> Self {
+    fn new(name: &str, about: &'a str) -> Self {
         Cli { app: clap_app!(
               (name) =>
               (about: about)
@@ -36,7 +55,7 @@ impl<'a, 'b> Cli<'a, 'b> {
               ), }
     }
 
-    pub fn add_base_packages_args(self) -> Self {
+    fn add_base_packages_args(self) -> Self {
         let app = self
             .app
             .arg(
@@ -78,7 +97,7 @@ impl<'a, 'b> Cli<'a, 'b> {
         Cli { app }
     }
 
-    pub fn add_builder_args(self) -> Self {
+    fn add_builder_args(self) -> Self {
         let app = self
             .app
             .arg(
@@ -129,7 +148,7 @@ impl<'a, 'b> Cli<'a, 'b> {
         Cli { app }
     }
 
-    pub fn add_tagging_args(self) -> Self {
+    fn add_tagging_args(self) -> Self {
         let app = self
             .app
             .arg(
@@ -183,7 +202,7 @@ impl<'a, 'b> Cli<'a, 'b> {
         Cli { app }
     }
 
-    pub fn add_publishing_args(self) -> Self {
+    fn add_publishing_args(self) -> Self {
         let app = self
             .app
             .arg(
@@ -248,7 +267,7 @@ impl<'a, 'b> Cli<'a, 'b> {
         Cli { app }
     }
 
-    pub fn add_pkg_ident_arg(self, options: PkgIdentArgOptions) -> Self {
+    fn add_pkg_ident_arg(self, options: PkgIdentArgOptions) -> Self {
         let help = if options.multiple {
             "One or more Habitat package identifiers (ex: acme/redis) and/or filepaths to a \
              Habitat Artifact (ex: /home/acme-redis-3.0.7-21120102031201-x86_64-linux.hart)"
@@ -267,7 +286,7 @@ impl<'a, 'b> Cli<'a, 'b> {
         Cli { app }
     }
 
-    pub fn add_memory_arg(self) -> Self {
+    fn add_memory_arg(self) -> Self {
         let app = self.app
                       .arg(Arg::with_name("MEMORY_LIMIT").value_name("MEMORY_LIMIT")
                                                          .long("memory")
@@ -278,7 +297,7 @@ impl<'a, 'b> Cli<'a, 'b> {
         Cli { app }
     }
 
-    pub fn add_base_image_arg(self) -> Self {
+    fn add_base_image_arg(self) -> Self {
         let app = self.app
                       .arg(Arg::with_name("BASE_IMAGE").value_name("BASE_IMAGE")
                                                        .long("base-image")
@@ -290,7 +309,7 @@ impl<'a, 'b> Cli<'a, 'b> {
         Cli { app }
     }
 
-    pub fn add_layer_arg(self) -> Self {
+    fn add_layer_arg(self) -> Self {
         let app =
             self.app
                 .arg(Arg::with_name("MULTI_LAYER").value_name("MULTI_LAYER")
